[WIP] Move test to sbtProj
